requirements.txt
discord.py~=2.0.1
asyncio~=3.4.3
typing~=3.6.6
requests~=2.28.1